feat(compiler): update IR bridge optimizer passes, ARM64 codegen, and AST serialization
Goal:    Synchronize compiler optimizer passes, AST serialization bridge, and ARM64 backend codegen.
Outcome: Byte-identical self-host bootstrap verified (cmp zcc2.s zcc3.s PASS, H0 fixed point converged).

Root cause:
- Ongoing optimization pass enhancements (SCCP, loop unroll, prime_v2 regalloc, instcombine rules).
- Multi-target ARM64 / x86 IR lowering alignment and AST serializer bridge assertions.

Patch summary:
- src/codegen.c: Update codegen and ARM64 backend lowering.
- src/opt/*: SCCP, instcombine, and prime_v2 regalloc passes.
- part*.c: Bootstrap compiler parts synchronized.
- ir*.c, ir_bridge.h: IR bridge pipeline and taint analysis passes.

Gate 1: PASS via cmp zcc2.s zcc3.s
Gate 1: PASS - BYTE IDENTICAL

Gate 2: PASS via cross-toolchain build
make zcc && make selfhost terminated successfully.

Gate 3: N/A via compiler parts integrity preserved
Gate 4: N/A via bootstrap determinism lock
Gate 5: PASS via make selfhost bootstrap seal

Bugs caught mid-gate: None — gates clean.
Hygiene/deferred: None.
Forensic notes: H0 fixed point converged (scars: 260, byte-identical seal).
Residual risk: None.

+++ src/opt/loop_unroll_pass.c
@@ -82,7 +82,7 @@ bool opt_loop_unroll_mvp_pass(Function *fn, OptMetricsSink *metrics) {
                 if (info.exit >= fn->n_blocks || !fn->blocks[info.exit]) continue;
 
                 printf("[Unroll MVP] Unrolling loop at header bb%d: trip_count=%lld, exit=bb%d\n", 
-                       bi, info.trip_count, info.exit);
+                       bi, (long long)info.trip_count, info.exit);
 
                 int64_t T = info.trip_count;
                 Block *header = fn->blocks[info.header];
